refactor(MP-112): Move user views to proper (access) folder.
